+++ src/functions/housingQueries.js
@@ -1,5 +1,5 @@
 import supabase from "../config/supabaseClient";
-import { computeTagsForHousing } from "../util/tagUtil";
+import { computeTagsForHousing } from "./util";
 import { getTagCountsForAllHousing, getTagCountsForHousing } from "./tagQueries";
 
 export const getAllHousing = async () => {
@@ -252,3 +252,19 @@ export const getAvgRatingByCategoryForHousing = async (id) => {
 
 	return data;
 };
+
+export const getReviewCountsForAllHousing = async () => {
+  let { data, error } = await supabase
+  .from('housing')
+  .select('id, reviews(count)')
+  .eq('reviews.status', 'approved');
+  if (error) {
+    console.log(`Error retrieving review counts`);
+    throw error;
+  }
+  const reviewCounts = {};
+  data.forEach((review) => {
+    reviewCounts[review.id] = review.reviews[0].count;
+  });
+  return reviewCounts;
+}
